Fix Readme
This patch fixes a number of errors on the readme page like typos, some
grammatical errors and invalid links. It also adds a very short build
guide since the linked documentation does not provide a guide and it's
always nice to not have to dig deep for a quick start.